wwfypc-django/requirements.txt
django
django-solo
gunicorn
mysqlclient
django-phonenumber-field
graphene-django>=2.0